requirements.txt
python-dotenv==1.2.1
urllib3==2.6.3
requests==2.32.5